+++ src/lib/app-provider/tx/storeCode.ts
@@ -40,7 +40,9 @@ export const useStoreCodeTx = (isMigrate: boolean) => {
 
       const message = composeStoreCodeMsg({
         sender: address,
-        wasmByteCode: await gzip(new Uint8Array(await wasmCode)),
+        wasmByteCode: new Uint8Array(
+          await gzip(new Uint8Array(await wasmCode))
+        ),
         permission,
         addresses,
       });