+++ src/lib/pages/custom-network/types.ts
@@ -1,7 +1,9 @@
 import type { ChainConfig } from "@alleslabs/shared";
+import { capitalize } from "lodash";
 import type { RefinementCtx } from "zod";
 import { z, ZodIssueCode } from "zod";
 
+import { getAccountBech32Lcd } from "lib/services/account/lcd";
 import {
   zAsset,
   zChainConfig,
@@ -14,6 +16,19 @@ import {
   zRegistry,
 } from "lib/types";
 
+import {
+  DEFAULT_BECH32_PREFIX,
+  DEFAULT_CUSTOM_MINITIA_NETWORK,
+  DEFAULT_GAS,
+  DEFAULT_GOV_CONFIG,
+  DEFAULT_MOVE_CONFIG,
+  DEFAULT_POOL_CONFIG,
+  DEFAULT_PUBLIC_PROJECT_CONFIG,
+  DEFAULT_SLIP44,
+  DEFAULT_WALLET_CONFIG,
+  DEFAULT_WASM_CONFIG,
+} from "./constant";
+
 const isAlphabetNumberAndSpecialCharacters = (str: string) =>
   /^[a-z0-9/:._-]+$/.test(str);
 
@@ -22,7 +37,6 @@ const mustBeAlphabetNumberAndSpecialCharacters =
 
 interface ValidateExistingChain {
   isChainIdExist: (chainId: string) => boolean;
-  isPrettyNameExist: (name: string) => boolean;
 }
 
 export enum VmType {
@@ -69,7 +83,7 @@ const networkDetailsFormValidator = (
   validateExistingChain: ValidateExistingChain
 ) => {
   const { prettyName, chainId, registryChainName } = val;
-  const { isChainIdExist, isPrettyNameExist } = validateExistingChain;
+  const { isChainIdExist } = validateExistingChain;
 
   if (prettyName === "")
     ctx.addIssue({
@@ -82,13 +96,6 @@ const networkDetailsFormValidator = (
     ctx.addIssue({
       code: ZodIssueCode.custom,
       message: `Minitia Name is too long. (${prettyName.length}/50)`,
-      path: ["prettyName"],
-    });
-
-  if (isPrettyNameExist(prettyName))
-    ctx.addIssue({
-      code: ZodIssueCode.custom,
-      message: "This name is already used. Please specify other name.",
       path: ["prettyName"],
     });
 
@@ -336,11 +343,9 @@ export type AddNetworkManualForm = z.infer<
 // MARK: AddNetworkManualChainConfigJson
 export const zAddNetworkManualChainConfigJson = ({
   isChainIdExist,
-  isPrettyNameExist,
 }: ValidateExistingChain) =>
   zAddNetworkManualForm({
     isChainIdExist,
-    isPrettyNameExist,
   }).transform<ChainConfig>(
     ({
       chainId,
@@ -362,9 +367,9 @@ export const zAddNetworkManualChainConfigJson = ({
       slip44,
       assets,
     }: AddNetworkManualForm) => ({
-      tier: "sequencer",
+      ...DEFAULT_CUSTOM_MINITIA_NETWORK,
+      wallets: DEFAULT_WALLET_CONFIG,
       chainId,
-      chain: "initia",
       registryChainName,
       prettyName,
       logo_URIs: {
@@ -372,26 +377,9 @@ export const zAddNetworkManualChainConfigJson = ({
       },
       lcd,
       rpc,
-      wallets: ["initia", "keplr"],
       features: {
-        faucet: {
-          enabled: false,
-        },
-        wasm:
-          vm.type === "wasm"
-            ? {
-                enabled: true,
-                storeCodeMaxFileSize: 1024 * 1024 * 2,
-                clearAdminGas: 1000000,
-              }
-            : { enabled: false },
-        move:
-          vm.type === "move"
-            ? {
-                enabled: true,
-                moduleMaxFileSize: 1_048_576,
-              }
-            : { enabled: false },
+        wasm: vm.type === "wasm" ? DEFAULT_WASM_CONFIG : { enabled: false },
+        move: vm.type === "move" ? DEFAULT_MOVE_CONFIG : { enabled: false },
         evm:
           vm.type === "evm"
             ? {
@@ -399,15 +387,9 @@ export const zAddNetworkManualChainConfigJson = ({
                 jsonRpc: vm.jsonRpc,
               }
             : { enabled: false },
-        pool: {
-          enabled: false,
-        },
-        publicProject: {
-          enabled: true,
-        },
-        gov: {
-          enabled: false,
-        },
+        pool: DEFAULT_POOL_CONFIG,
+        publicProject: DEFAULT_PUBLIC_PROJECT_CONFIG,
+        gov: DEFAULT_GOV_CONFIG,
         nft: {
           enabled: vm.type === "move",
         },
@@ -416,11 +398,6 @@ export const zAddNetworkManualChainConfigJson = ({
         gasAdjustment,
         maxGasLimit,
       },
-      extra: {
-        isValidatorExternalLink: null,
-        layer: "2",
-      },
-      network_type: "local",
       fees: {
         fee_tokens: [
           {
@@ -477,15 +454,80 @@ export const zAddNetworkJsonChainConfigJson = zChainConfig
   })
   .transform<ChainConfig>((val) => ({
     ...val,
-    tier: "sequencer",
-    chain: "initia",
-    extra: {
-      isValidatorExternalLink: null,
-      layer: "2",
-    },
-    network_type: "local",
+    ...DEFAULT_CUSTOM_MINITIA_NETWORK,
   }));
 
 export type AddNetworkJsonChainConfigJson = z.infer<
   typeof zAddNetworkJsonChainConfigJson
 >;
+
+export const zAddNetworkLinkChainConfigJson = z
+  .object({
+    chainId: z.string().min(1, "Chain ID cannot be empty"),
+    lcd: zHttpsUrl,
+    rpc: zHttpsUrl,
+    jsonRpc: zHttpsUrl.optional(),
+    vm: z.nativeEnum(VmType),
+    minGasPrice: z.number(),
+    denom: z.string(),
+  })
+  .superRefine((val, ctx) => {
+    if (val.vm === VmType.EVM && !val.jsonRpc) {
+      ctx.addIssue({
+        code: ZodIssueCode.custom,
+        message: "jsonRpc is required when vm is evm",
+        path: ["jsonRpc"],
+      });
+    }
+  })
+  .transform<ChainConfig>(async (val) => {
+    const bech32Prefix = await getAccountBech32Lcd(val.lcd)
+      .then((res) => res.bech32Prefix)
+      .catch(() => DEFAULT_BECH32_PREFIX);
+
+    return {
+      ...DEFAULT_CUSTOM_MINITIA_NETWORK,
+      chainId: val.chainId,
+      registryChainName: val.chainId,
+      prettyName: capitalize(val.chainId),
+      wallets: DEFAULT_WALLET_CONFIG,
+      lcd: val.lcd,
+      rpc: val.rpc,
+      features: {
+        wasm: val.vm === VmType.WASM ? DEFAULT_WASM_CONFIG : { enabled: false },
+        move: val.vm === VmType.MOVE ? DEFAULT_MOVE_CONFIG : { enabled: false },
+        evm:
+          val.vm === VmType.EVM
+            ? { enabled: true, jsonRpc: val.jsonRpc! } // jsonRpc is required when vm is evm
+            : { enabled: false },
+        pool: DEFAULT_POOL_CONFIG,
+        publicProject: DEFAULT_PUBLIC_PROJECT_CONFIG,
+        gov: DEFAULT_GOV_CONFIG,
+        nft: {
+          enabled: val.vm === VmType.MOVE,
+        },
+      },
+      gas: {
+        gasAdjustment: DEFAULT_GAS.gasAdjustment,
+        maxGasLimit: DEFAULT_GAS.maxGasLimit,
+      },
+      fees: {
+        fee_tokens: [
+          {
+            denom: val.denom,
+            fixed_min_gas_price: val.minGasPrice,
+            low_gas_price: val.minGasPrice,
+            average_gas_price: val.minGasPrice,
+          },
+        ],
+      },
+      registry: {
+        bech32_prefix: bech32Prefix,
+        slip44: DEFAULT_SLIP44,
+        staking: {
+          staking_tokens: [],
+        },
+        assets: [],
+      },
+    };
+  });